Extract deliverNotifications into shared function between cron and request
