+++ functions/src/notifications/index.ts
@@ -3,9 +3,8 @@ import { publishNotifications } from "./publishNotifications"
 import { populateBillHistoryNotificationEvents } from "./populateBillHistoryNotificationEvents"
 import { populateTestimonySubmissionNotificationEvents } from "./populateTestimonySubmissionNotificationEvents"
 import { cleanupNotifications } from "./cleanupNotifications"
-import { deliverNotifications } from "./deliverNotifications"
+import { deliverNotifications, httpsDeliverNotifications } from "./deliverNotifications"
 import { httpsPublishNotifications } from "./httpsPublishNotifications"
-import { httpsDeliverNotifications } from "./httpsDeliverNotifications"
 import { httpsCleanupNotifications } from "./httpsCleanupNotifications"
 import { updateUserNotificationFrequency } from "./updateUserNotificationFrequency"
 import { updateNextDigestAt } from "./updateNextDigestAt"
